Add dev-automation pipe and DX11 tooling
Introduce a dev-automation channel and a suite of developer tooling to support automated testing, DX11 shader builds, and improved capture/crash workflows.

Highlights:
- Add engine/devautomation.{cpp,h} and wire into host to accept local named-pipe commands (-devautomation) so test tools can drive the engine safely from the host thread.
- New tf2ctl.py: Python test client to launch TF2, send console commands over the pipe, wait on console.log patterns, take in-engine screenshots and manage graceful kills.
- Add PowerShell launch wrappers for HL2/TF2 (renderdoc and cdb capture flows) and runtime .hl2runtime.bat for local dev paths.
- Add DX11 shader build driver (devtools/build_shaders_dx11.py) and many shader/renderer DX11 support stubs and HLSL assets.
- Various runtime improvements: PIXEVENT instrumentation for render grouping, TF GC auth-ticket watchdog to recover stalled Steam calls, steamless-safe checks in loadout UI, and a dev_dismiss_ui console command to close menus for automation.
- Waf build script additions and small .gitignore update for captures.

These changes enable reliable automated visual testing, crash capture, and DX11 shader packaging for development workflows.

+++ hammer/material.cpp
@@ -434,7 +434,7 @@ CMaterial::~CMaterial(void)
 // Finds all .VMT files in a particular directory
 //-----------------------------------------------------------------------------
 bool CMaterial::LoadMaterialsInDirectory( char const* pDirectoryName, int nDirectoryNameLen,
-						IMaterialEnumerator *pEnum, int nContext, int nFlags )
+						IMaterialEnumerator *pEnum, intp nContext, int nFlags )
 {
 	//Assert( Q_strnicmp( pDirectoryName, "materials", 9 ) == 0 );
 
@@ -489,7 +489,7 @@ bool CMaterial::LoadMaterialsInDirectory( char const* pDirectoryName, int nDirec
 // that we can load up at a later time 
 //-----------------------------------------------------------------------------
 bool CMaterial::InitDirectoryRecursive( char const* pDirectoryName, 
-						IMaterialEnumerator *pEnum, int nContext, int nFlags )
+						IMaterialEnumerator *pEnum, intp nContext, int nFlags )
 {
 	// Make sure this is an ok directory, otherwise don't bother
 	if (ShouldSkipMaterial( pDirectoryName + MATERIAL_PREFIX_LEN, nFlags ))
@@ -539,7 +539,7 @@ bool CMaterial::InitDirectoryRecursive( char const* pDirectoryName,
 // It only finds their names so we can generate shell materials for them
 // that we can load up at a later time 
 //-----------------------------------------------------------------------------
-void CMaterial::EnumerateMaterials( IMaterialEnumerator *pEnum, const char *szRoot, int nContext, int nFlags )
+void CMaterial::EnumerateMaterials( IMaterialEnumerator *pEnum, const char *szRoot, intp nContext, int nFlags )
 {
 	InitDirectoryRecursive( szRoot, pEnum, nContext, nFlags );
 }
